requirements.txt
Django==1.11.7
dj-database-url==0.4.2
gunicorn==19.7.1
psycopg2==2.7.3.2
whitenoise==3.3.1
python-decouple==3.1
django-oscar==1.5.1
django_compressor==2.2
pysolr==3.7.0
dj-static==0.0.6
static3==0.7.0
brotlipy==0.7.0
cffi==1.11.2
pycparser==2.18
django-localflavor==1.6.2
django-debug-toolbar==1.9.1
django-extensions==1.9.8		
django-environ==0.4.4




